Fixing cookie
Before cookie was saved but not rendered properly
